# app-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Notatek",
        "description": "Dokumentacja API dla systemu notatek",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Pobieranie profilu zalogowanego użytkownika",
                "operationId": "d4bdd8cfcb7e0150221c67124f47874d",
                "responses": {
                    "200": {
                        "description": "Dane użytkownika"
                    },
                    "401": {
                        "description": "Nieautoryzowany dostęp"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Aktualizacja profilu użytkownika",
                "operationId": "7e008f8cf5f01a905c53edd0126184bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "newuser"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newsecurepassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil zaktualizowany pomyślnie"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/dashboard": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Wyświetlanie dashboardu użytkownika",
                "operationId": "a1c7c0f70150a745a2eeb2a4d7ba890c",
                "responses": {
                    "200": {
                        "description": "Dane użytkownika i jego notatki"
                    },
                    "401": {
                        "description": "Nieautoryzowany dostęp"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/me/profile-picture": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Aktualizacja zdjęcia profilowego użytkownika",
                "operationId": "048354cf6daf0fba6f8b4dd3a7931f34",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "profile_picture": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zdjęcie profilowe zaktualizowane pomyślnie"
                    },
                    "400": {
                        "description": "Nieprawidłowy format pliku"
                    },
                    "500": {
                        "description": "Błąd serwera"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Wpisz token w formacie (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}